requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
asyncpg>=0.29.0
httpx>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
prefect>=2.14.0
